refactoring OperationsDataManager
